Refetch channel info
Some users with dozens of accounts have seen an issue with Eve clients
where the character name information is not available in the log  at the
time messages start appearing in game logs. This wasn't thought to be
possible except in rare cases such as creating a new character. Since it
seems it can happen when running dozens of clients and/or using isboxer,
logic needs to be added to handle this. So each time a LogReader is
polled, it will recheck channel+character name if either is still blank.